api/requirements.txt
redis==7.3.0
fastapi==0.135.1
python-dotenv==1.1.1
pydantic==2.11.9
uvicorn==0.35.0
websockets==15.0.1
pandas_market_calendars==5.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
